week07/airflow_dbx/requirements.txt
apache-airflow>=2.6.0
apache-airflow-providers-databricks>=6.0.0
